masking: run SAM2 locally on Apple-silicon (mps)
Enable the SAM2 video tracker on an M-series Mac GPU for local mask QC:
- resolve_device(): cuda configs fall through to mps, then cpu, instead of
  silently dropping to cpu on machines without CUDA.
- _load_img_as_tensor(): cast frames to float32 — uint8/255.0 promotes to
  float64 under numpy and the async-loader path can't .to('mps') float64.
- run_on_video.py: set PYTORCH_ENABLE_MPS_FALLBACK=1 before torch import so
  the few ops MPS lacks fall back to CPU rather than crashing.

Verified: endo18 checkpoint tracks a clicked instrument across a 449-frame
test clip at ~10 fps on an M3 Max.

+++ sam2/utils/misc.py
@@ -96,7 +96,9 @@ def _load_img_as_tensor(img_path, image_size):
         img_np = img_np / 255.0
     else:
         raise RuntimeError(f"Unknown image dtype: {img_np.dtype} on {img_path}")
-    img = torch.from_numpy(img_np).permute(2, 0, 1)
+    # uint8 / 255.0 promotes to float64 under numpy; cast back to float32 so
+    # the async-loader path stays MPS-safe (MPS rejects float64 tensors).
+    img = torch.from_numpy(img_np).permute(2, 0, 1).float()
     video_width, video_height = img_pil.size  # the original video size
     return img, video_height, video_width
